+++ CHANGELOG.md
@@ -5,8 +5,8 @@
 ### New features
 - Implemented report views. See [Report views](./README.md#report-views) for details.
 - Reworked [Delta-Coverage GitHub Action](actions/delta-coverage-report/README.md). 
-  Now it publishes Delta-Coverage report using GitHub Check API.
-  PR comment now contains only links to the check runs.
+  - Now it publishes Delta-Coverage report using GitHub Check API.
+  - PR comment now contains only links to the check runs and short summary.
 
 ### Breaking changes
 - Id of the plugin was changed to `io.github.gw-kit.delta-coverage`.